Fixed a bug with adding a new charge
